# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Vook Frontend API
  version: 2.0.0
  description: Resource-oriented contract consumed by the standalone Vook frontend.
servers: [{ url: /api/v2 }]
security: [{ sessionCookie: [] }]
paths:
...
  /role-permissions:
    get: { operationId: listRolePermissions, responses: { '200': { $ref: '#/components/responses/Object' } } }
    put: { operationId: updateRolePermissions, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
  /role-definitions:
    get: { operationId: listRoleDefinitions, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: createRoleDefinition, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /role-definitions/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put: { operationId: updateRoleDefinition, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
    delete: { operationId: deleteRoleDefinition, responses: { '200': { $ref: '#/components/responses/Empty' } } }
  /role-assignments:
    get: { operationId: listRoleAssignments, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: createRoleAssignment, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /role-assignments/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    delete: { operationId: deleteRoleAssignment, responses: { '200': { $ref: '#/components/responses/Empty' } } }
...
  /workflows/{type}:
    parameters: [{ name: type, in: path, required: true, schema: { type: string } }]
    put: { operationId: updateWorkflow, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: vook_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  requestBodies:
    Object: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  responses:
    Empty: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/EmptyEnvelope' } } } }
    Object: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/ObjectEnvelope' } } } }
...
  schemas:
...
    ListMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1 }, total: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ObjectEnvelope: { type: object, required: [data], properties: { data: {}, meta: { $ref: '#/components/schemas/ListMeta' } } }
    EmptyEnvelope: { type: object, required: [data], properties: { data: { type: 'null' } } }
...